requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
Cython==0.29.13
graphviz==0.13
grpcio==1.23.0
grpcio-tools==1.23.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
msgpack==0.6.1
mypy-protobuf==1.15
numpy==1.17.1
objgraph==3.4.1
opencv-python==4.1.1.26
protobuf==3.9.1
pycocotools==2.0.0
Pympler==0.7
pyparsing==2.4.2
PyQt5==5.13.0
PyQt5-sip==4.19.18
PyQt5-stubs==5.13.0.1
python-dateutil==2.8.0
requests==2.22.0
retrying==1.3.3
serious==1.0.0.dev20
six==1.12.0
urllib3==1.25.3
